Aims to fix fermentation starting issues.

+++ database/migrations/2021_11_22_000007_create_fermentations_table.php
@@ -19,7 +19,7 @@ class CreateFermentationsTable extends Migration
             $table->foreignId("brewed_by")->constrained("users");
             $table->foreignId("equipment_id")->nullable()->comment("Equipment Table is not implemented in MVP.")->constrained("equipment");
             $table->timestamp("started_at")->nullable();
-            $table->float("og")->nullable();
+            $table->decimal("og", 10, 4)->nullable();
             $table->boolean("completed")->default(false);
             $table->timestamps();
         });
